Removed redundant universal code
